[Modify] Book, recommand db 컬럼 수정

+++ server/src/main/java/com/emosaac/server/domain/book/Score.java
@@ -29,7 +29,7 @@ public class Score{
     @JoinColumn(name = "BOOK_NO")
     private Book book;
 
-    private Long score;
+    private double score;
 
 
 }
